src/requirements.txt
files==1.1.1
Pillow==9.2.0
